Use enum value for registers with singleton field
Registers with singleton fields are registers which:

    - Define a single field (bitfield)
    - This field has the same size as the register

When this is the case and the field has an enum, the enum is used as a
value type to read/write directly to the register. Bitfield accessors
are not generated.

This required adding a `size` annotation to all enum values so that
their size matches their parent register size.