Generate new peripheral types when prepend/append differs
SVD tags prependToName/appendToName instruct svd2nim to add a prefix
or suffix to child register member names.

This resulted in bad codegen when svd2nim tried to reuse a peripheral
object type for a derived peripheral, but the derived peripheral
redefined prependToName or appendToName, leading to different member
names and therefore incompatible types.

This manifested notably in the PORT_IOBUS peripheral of
ATSAMD21G18A.svd, which is derived from PORT but defines
prependToName="PORT_IOBUS".

To fix this, we generate a new type when the peripheral is derived but
has a different prependToName or appendToName (and the corresponding
append/prepend is not ignored by a CLI codegen option.)